% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t2df.R
\name{addE}
\alias{addE}
\title{Add an edge or a list of edges to the data frame for edges.}
\usage{
addE(df, from, to)
}
\arguments{
\item{df}{Data frame for edges.}

\item{from}{Integer (numerical identifier of vertex).}

\item{to}{Vector of integers (numerical identifiers of kids of vertex).}
}
\value{
A Data frame for edges.
}
\description{
Adds the edges of the kids of a node to the edge list.
}
\examples{
df<-addE(newE(), 1, c(2, 3, 4))
print(df) 

}
\seealso{
Other Data frames for igraph: 
\code{\link{addV}()},
\code{\link{newE}()},
\code{\link{newV}()}
}
\concept{Data frames for igraph}
