% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t2df.R
\name{addV}
\alias{addV}
\title{Add a vertex to the data frame for vertices.}
\usage{
addV(df, id, name)
}
\arguments{
\item{df}{Data frame for vertices.}

\item{id}{Integer (numerical identifier of vertex).}

\item{name}{Name of grammar symbol of of vertex.}
}
\value{
A Data frame for vertices.
}
\description{
Add a vertex to the data frame for vertices.
}
\examples{
df<-addV(newV(), 1, "<fe>")
print(df) 

}
\seealso{
Other Data frames for igraph: 
\code{\link{addE}()},
\code{\link{newE}()},
\code{\link{newV}()}
}
\concept{Data frames for igraph}
