% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derivationTrees.R
\name{booleanGrammar}
\alias{booleanGrammar}
\title{A constant function which returns the BNF (Backus-Naur Form) 
of a context-free grammar for the XOR problem.}
\usage{
booleanGrammar()
}
\value{
A named list with elements \code{$filename} and  \code{$BNF} 
        representing the grammar of a boolean grammar with two variables and
        the boolean functions \code{AND}, \code{OR}, and \code{NOT}.
}
\description{
A constant function which returns the BNF (Backus-Naur Form) 
of a context-free grammar for the XOR problem.
}
\details{
Imported from package xegaBNF for use in examples.
}
\examples{
booleanGrammar()
}
\seealso{
Other Grammar: 
\code{\link{compileBNF}()}
}
\concept{Grammar}
