% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geTree.R
\name{chooseRulek}
\alias{chooseRulek}
\title{Selects k-th production rule index from a vector of production rules.}
\usage{
chooseRulek(riv, k)
}
\arguments{
\item{riv}{Vector of production rules indices for 
a non-terminal symbol.}

\item{k}{Integer.}
}
\value{
The index of the production rule.
}
\description{
\code{chooseRulek()} selects the k-th production rule index 
     from the vector of production rule indices 
     in the \code{g$PT$LHS$} for a non-terminal symbol.
}
\examples{
chooseRulek(c(7, 8, 9), 9)
chooseRulek(as.vector(1), 9)
}
\concept{Choice}
