% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derivationTrees.R
\name{decodeDT}
\alias{decodeDT}
\title{Decodes a derivation tree into a program.}
\usage{
decodeDT(tree, ST)
}
\arguments{
\item{tree}{Derivation tree.}

\item{ST}{Symbol table.}
}
\value{
Program
}
\description{
The program may contain non-terminal symbols
             and its evaluation may fail.
}
\examples{
g<-compileBNF(booleanGrammar())
t1<-generateDerivationTree(sym=g$Start,sample(100, 10, replace=TRUE), G=g)
decodeDT(t1$tree, g$ST) 

}
\seealso{
Other Decoder: 
\code{\link{decodeCDT}()},
\code{\link{decodeDTsym}()},
\code{\link{decodeTree}()},
\code{\link{leavesIncompleteDT}()}
}
\concept{Decoder}
