% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printDerivations.R
\name{decodeSymVec}
\alias{decodeSymVec}
\title{Decodes a vector of symbols.}
\usage{
decodeSymVec(v, ST)
}
\arguments{
\item{v}{Vector of symbols.}

\item{ST}{Symbol table.}
}
\value{
A program.
}
\description{
Decodes a vector of symbols.
}
\examples{
g<-compileBNF(booleanGrammar())
a<-randomDerivationTree(g$Start, g)
r<-treeRoot(a)
decodeSymVec(r, g$ST)
c<-unlist(lapply(treeChildren(a), FUN=treeRoot))
decodeSymVec(c, g$ST)

}
