% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t2df.R
\name{newE}
\alias{newE}
\title{Returns empty data frame for edges.}
\usage{
newE()
}
\value{
Data frame.
}
\description{
Data frame with columns 
             \code{from} root (numerical identifier) and 
             \code{to} kid (numerical identifier).
}
\examples{
df<-newE()
print(df) 

}
\seealso{
Other Data frames for igraph: 
\code{\link{addE}()},
\code{\link{addV}()},
\code{\link{newV}()}
}
\concept{Data frames for igraph}
