% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t2df.R
\name{newV}
\alias{newV}
\title{Returns empty data frame for vertices.}
\usage{
newV()
}
\value{
Data frame.
}
\description{
Data frame with columns 
             \code{id} (numerical identifier) and 
             \code{name} (symbol of the grammar).
}
\examples{
df<-newV()
print(df) 

}
\seealso{
Other Data frames for igraph: 
\code{\link{addE}()},
\code{\link{addV}()},
\code{\link{newE}()}
}
\concept{Data frames for igraph}
