% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printDerivations.R
\name{printDerivations}
\alias{printDerivations}
\title{Print derivations.}
\usage{
printDerivations(tree, G, verbose = FALSE)
}
\arguments{
\item{tree}{Derivation tree.}

\item{G}{The context-free grammar.}

\item{verbose}{If TRUE, the list of derivations is printed.
Default: FALSE.}
}
\value{
A list of derivations.
}
\description{
A depth-first left-to-right tree traversal
              without recursion.
}
\details{
Works with complete and incomplete derivation trees.
}
\examples{

g<-compileBNF(booleanGrammar())
a<-randomDerivationTree(g$Start, g)
l<-printDerivations(a, g, verbose=TRUE) 

}
\concept{Diagnostics}
