% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derivationTrees.R
\name{rndPartition}
\alias{rndPartition}
\title{Randomly partitions n in k parts.}
\usage{
rndPartition(n, k)
}
\arguments{
\item{n}{The integer to divide.}

\item{k}{Number of parts.}
}
\value{
The integer partition of n in k parts.
}
\description{
Sampling a partition is a two-step process:
  \enumerate{
  \item The k parts of the partion are sampled in the loop.
        This implies that the first partition p is a random number 
        between 1 and 1+n-k. The next partition is sampled from 
        1 to 1+n-k-p. 
  \item We permute the partitions.      
  }
}
\examples{
 rndPartition(10, 4)
}
\concept{Unused}
