% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derivationTrees.R
\name{substituteSymbol}
\alias{substituteSymbol}
\title{Codes the substitution of a non-terminal symbol by the symbols 
derived by a production rule as a nested list.}
\usage{
substituteSymbol(rindex, PT)
}
\arguments{
\item{rindex}{Rule index.}

\item{PT}{Production table.}
}
\value{
2-element list.
}
\description{
\code{substituteSymbol()} 
generates a nested list with the non-terminal symbol as the root 
(first list element) and the derived symbols as the second list element.
}
\examples{
g<-compileBNF(booleanGrammar())
substituteSymbol(3, g$PT)

}
\seealso{
Other Generate Derivation Tree: 
\code{\link{generateCDT}()},
\code{\link{generateDerivationTree}()},
\code{\link{randomDerivationTree}()},
\code{\link{rndsub}()},
\code{\link{rndsubk}()}
}
\concept{Generate Derivation Tree}
