% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geTree.R
\name{testGenerateDerivationTree}
\alias{testGenerateDerivationTree}
\title{Generate, decode, and show \code{times} derivation trees from random 
integer vectors for grammar BNF on the console.}
\usage{
testGenerateDerivationTree(times, BNF, verbose = TRUE)
}
\arguments{
\item{times}{Number of derivation trees which should be generated.}

\item{BNF}{BNF.}

\item{verbose}{Boolean. If TRUE (default) , print decoded derivation tree on console.}
}
\value{
Number of complete derivation trees generated.
}
\description{
Generate, decode, and show \code{times} derivation trees from random 
integer vectors for grammar BNF on the console.
}
\examples{
testGenerateDerivationTree(5, BNF=booleanGrammar())
}
\concept{Tests}
