% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derivationTrees.R
\name{treeChildren}
\alias{treeChildren}
\title{Returns the children of a derivation tree.}
\usage{
treeChildren(tree)
}
\arguments{
\item{tree}{Derivation tree.}
}
\value{
The children of a derivation tree (a list of derivation trees).
}
\description{
\code{treeChildren()} returns the children of a derivation tree
         represented as a list of derivation trees.
}
\examples{
g<-compileBNF(booleanGrammar())
a<-randomDerivationTree(g$Start, g)
treeChildren(a) 

}
\seealso{
Other Access Tree Parts: 
\code{\link{filterANL}()},
\code{\link{filterANLid}()},
\code{\link{treeANL}()},
\code{\link{treeRoot}()}
}
\concept{Access Tree Parts}
