% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derivationTrees.R
\name{treeRoot}
\alias{treeRoot}
\title{Returns the root of a derivation tree.}
\usage{
treeRoot(tree)
}
\arguments{
\item{tree}{Derivation tree.}
}
\value{
Root of a derivation tree.
}
\description{
\code{treeRoot()} returns the root of a derivation tree.
}
\examples{
g<-compileBNF(booleanGrammar())
a<-randomDerivationTree(g$Start, g)
treeRoot(a) 

}
\seealso{
Other Access Tree Parts: 
\code{\link{filterANLid}()},
\code{\link{filterANL}()},
\code{\link{treeANL}()},
\code{\link{treeChildren}()}
}
\concept{Access Tree Parts}
